AER-418 ADMS diurnal variation IMAER, pt 2 (#94)
* Added diurnal variation to ADMS objects in IMAER
Also changed the values from integer to double
* Roundtrip test for ADMS source characteristics with diurnal variation
* Went from average value of 10000 to 1
Discussed with other developers, 1 makes more sense if we're using doubles anyway, else we'd have to multiply with a factor in different locations.
Changed the check in GML2Definitions to use an epsilon instead: that way we should be OK with having inconsistencies due to floating point calculations.
Removed the 4_0 example, as it wasn't really used in that version anyway, and it's a bit hard to go from integers to doubles and still have a meaningful profile that actually tests something...

+++ source/imaer-gml/src/main/java/nl/overheid/aerius/gml/base/characteristics/IsGmlCustomDiurnalVariation.java
@@ -26,6 +26,6 @@ public interface IsGmlCustomDiurnalVariation {
 
   String getCustomType();
 
-  List<Integer> getValues();
+  List<Double> getValues();
 
 }
